fix(watcher): kill process group + bound stderr drain in kill paths
Fixes #116. When watch_streaming hit a kill path (idle timeout, cost
ceiling, or stuck-loop detection) it called child.kill() on the direct
child only, then awaited the stderr capture handle. If the agent had
spawned descendants in the same process group, those descendants kept
the stderr pipe open, the handle await blocked forever, watch_streaming
never returned, and update_task_completion was never called — leaving
the task status as Running indefinitely. aid watch --quiet would block
until an external aid stop flipped the status.

Now each kill path calls force_kill_process_group(child) before
child.kill(), and every stderr handle await is wrapped in a 2s timeout
via drain_stderr_capture. Also extracted the pgroup helpers into a
shared crate::process_group module and deduped the copies in
run_agent.rs and run_process.rs.
